Moves the test case to HadoopFsRelationTest
